Move scroll component to Core
